Moved all the source files to src folder
